react freindly components setup
