+++ src/components/Play/games/GymRoomScene.js
@@ -1,0 +1,13 @@
+import Phaser from "phaser";
+
+export default class GymRoom extends Phaser.Scene {
+    constructor() {
+        super({ key: 'gym-main-room' });
+    }
+
+    preload() { }
+
+    create() { }
+
+    update(time, delta) { }
+}
